Write database models, HTTP API and tests for backend (#2)
We shouldn't have to make too many changes to the backend after this,
except to serve the HTML and JS from the ui app. I may add some more
unit tests for the error handling in the HTTP API.

I wrote the API and models with the full application in mind, not just
the barebones initial implementation, so we won't have to go back and
change things later.